--- alembic/versions/3dc125c52fee_.py
@@ -1,38 +1,0 @@
-"""empty message
-
-Revision ID: 3dc125c52fee
-Revises: fc3f0137f538
-Create Date: 2023-02-26 15:51:11.555372
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '3dc125c52fee'
-down_revision = 'fc3f0137f538'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('otp_authentication',
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
-    sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=False),
-    sa.Column('tries', sa.SmallInteger(), nullable=False),
-    sa.Column('hashed_cors_token', sa.String(length=97), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_otp_authentication_id'), 'otp_authentication', ['id'], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_otp_authentication_id'), table_name='otp_authentication')
-    op.drop_table('otp_authentication')
-    # ### end Alembic commands ###
