+++ alembic/versions/702abfef734b_.py
@@ -1,0 +1,46 @@
+"""empty message
+
+Revision ID: 702abfef734b
+Revises: c37896fb58cc
+Create Date: 2023-02-17 22:00:57.568000
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision = '702abfef734b'
+down_revision = 'c37896fb58cc'
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('cron_report',
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_cron_report_id'), 'cron_report', ['id'], unique=True)
+    op.create_table('cron_report_data',
+    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
+    sa.Column('encrypted_report', sa.String(length=5000), nullable=False),
+    sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=True),
+    sa.Column('report_id', postgresql.UUID(as_uuid=True), nullable=True),
+    sa.ForeignKeyConstraint(['report_id'], ['cron_report.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_cron_report_data_id'), 'cron_report_data', ['id'], unique=True)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_cron_report_data_id'), table_name='cron_report_data')
+    op.drop_table('cron_report_data')
+    op.drop_index(op.f('ix_cron_report_id'), table_name='cron_report')
+    op.drop_table('cron_report')
+    # ### end Alembic commands ###
