+++ alembic/versions/7e9ad1bb809e_.py
@@ -1,8 +1,8 @@
 """empty message
 
-Revision ID: 7f6c999e53ce
-Revises: 3dc125c52fee
-Create Date: 2023-03-01 12:30:49.353603
+Revision ID: 7e9ad1bb809e
+Revises: a01f82635acf
+Create Date: 2023-03-01 21:23:13.505233
 
 """
 from alembic import op
@@ -10,14 +10,25 @@ import sqlalchemy as sa
 from sqlalchemy.dialects import postgresql
 
 # revision identifiers, used by Alembic.
-revision = '7f6c999e53ce'
-down_revision = '3dc125c52fee'
+revision = '7e9ad1bb809e'
+down_revision = 'a01f82635acf'
 branch_labels = None
 depends_on = None
 
 
 def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('user_otp',
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
+    sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=False),
+    sa.Column('secret', sa.String(), nullable=False),
+    sa.Column('status', sa.Enum('AVAILABLE', 'AWAITING_VERIFICATION', name='otpstatustype'), nullable=True),
+    sa.Column('hashed_recovery_codes', sa.ARRAY(sa.String()), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_user_otp_id'), 'user_otp', ['id'], unique=True)
     op.drop_index('ix_image_proxy_id', table_name='image_proxy')
     op.drop_table('image_proxy')
     # ### end Alembic commands ###
@@ -35,4 +46,6 @@ def downgrade() -> None:
     sa.PrimaryKeyConstraint('id', name='image_proxy_pkey')
     )
     op.create_index('ix_image_proxy_id', 'image_proxy', ['id'], unique=False)
+    op.drop_index(op.f('ix_user_otp_id'), table_name='user_otp')
+    op.drop_table('user_otp')
     # ### end Alembic commands ###
