+++ alembic/versions/c37896fb58cc_.py
@@ -1,8 +1,8 @@
 """empty message
 
-Revision ID: 3abacd70e563
+Revision ID: c37896fb58cc
 Revises: 
-Create Date: 2023-02-25 21:13:11.670518
+Create Date: 2023-02-15 12:25:09.538104
 
 """
 from alembic import op
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 from sqlalchemy.dialects import postgresql
 
 # revision identifiers, used by Alembic.
-revision = '3abacd70e563'
+revision = 'c37896fb58cc'
 down_revision = None
 branch_labels = None
 depends_on = None
@@ -18,12 +18,6 @@ depends_on = None
 
 def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('cron_report',
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_cron_report_id'), 'cron_report', ['id'], unique=True)
     op.create_table('deleted_email_alias',
     sa.Column('created_at', sa.DateTime(), nullable=False),
     sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
@@ -80,16 +74,6 @@ def upgrade() -> None:
     sa.PrimaryKeyConstraint('id')
     )
     op.create_index(op.f('ix_user_id'), 'user', ['id'], unique=True)
-    op.create_table('cron_report_data',
-    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
-    sa.Column('encrypted_report', sa.String(length=5000), nullable=False),
-    sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=True),
-    sa.Column('report_id', postgresql.UUID(as_uuid=True), nullable=True),
-    sa.ForeignKeyConstraint(['report_id'], ['cron_report.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_cron_report_data_id'), 'cron_report_data', ['id'], unique=True)
     op.create_table('email',
     sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
     sa.Column('address', sa.String(length=400), nullable=False),
@@ -150,16 +134,6 @@ def upgrade() -> None:
     sa.PrimaryKeyConstraint('id')
     )
     op.create_index(op.f('ix_m2m_reserved_alias_user_id'), 'm2m_reserved_alias_user', ['id'], unique=True)
-    op.create_table('user_otp',
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
-    sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=False),
-    sa.Column('secret', sa.String(), nullable=False),
-    sa.Column('status', sa.Enum('AVAILABLE', 'AWAITING_VERIFICATION', name='otpstatustype'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_user_otp_id'), 'user_otp', ['id'], unique=True)
     op.create_table('user_preferences',
     sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
     sa.Column('user_id', postgresql.UUID(as_uuid=True), nullable=True),
@@ -192,8 +166,6 @@ def downgrade() -> None:
     op.drop_table('image_proxy')
     op.drop_index(op.f('ix_user_preferences_id'), table_name='user_preferences')
     op.drop_table('user_preferences')
-    op.drop_index(op.f('ix_user_otp_id'), table_name='user_otp')
-    op.drop_table('user_otp')
     op.drop_index(op.f('ix_m2m_reserved_alias_user_id'), table_name='m2m_reserved_alias_user')
     op.drop_table('m2m_reserved_alias_user')
     op.drop_index(op.f('ix_email_report_id'), table_name='email_report')
@@ -207,8 +179,6 @@ def downgrade() -> None:
     op.drop_index(op.f('ix_email_id'), table_name='email')
     op.drop_index(op.f('ix_email_address'), table_name='email')
     op.drop_table('email')
-    op.drop_index(op.f('ix_cron_report_data_id'), table_name='cron_report_data')
-    op.drop_table('cron_report_data')
     op.drop_index(op.f('ix_user_id'), table_name='user')
     op.drop_table('user')
     op.drop_index(op.f('ix_server_statistics_id'), table_name='server_statistics')
@@ -221,6 +191,4 @@ def downgrade() -> None:
     op.drop_table('global_settings')
     op.drop_index(op.f('ix_deleted_email_alias_id'), table_name='deleted_email_alias')
     op.drop_table('deleted_email_alias')
-    op.drop_index(op.f('ix_cron_report_id'), table_name='cron_report')
-    op.drop_table('cron_report')
     # ### end Alembic commands ###
