--- alembic/versions/fc3f0137f538_.py
@@ -1,28 +1,0 @@
-"""empty message
-
-Revision ID: fc3f0137f538
-Revises: 3abacd70e563
-Create Date: 2023-02-26 11:05:45.074081
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'fc3f0137f538'
-down_revision = '3abacd70e563'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user_otp', sa.Column('hashed_recovery_codes', sa.ARRAY(sa.String()), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user_otp', 'hashed_recovery_codes')
-    # ### end Alembic commands ###
